requirements.txt
alembic==1.8.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
Mako==1.2.3
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.5.0
psycopg2==2.9.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
urllib3==1.26.12
xmltodict==0.13.0
zipp==3.9.0
